% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{as.binary}
\alias{as.binary}
\alias{is.binary}
\alias{try_binary}
\title{Work with Binary Columns}
\usage{
as.binary(x)

is.binary(x)

try_binary(x)
}
\arguments{
\item{x}{value(s) to convert to \code{binary}}
}
\description{
Transform \link{logical}s and \link{numeric}s to a new class 'binary'. The function \code{\link[=try_binary]{try_binary()}} only coerces to \code{binary} if the input is numeric and consists of only zeroes and ones.
}
\examples{
as.binary(c(TRUE, FALSE))

as.binary(c(1, 0))
as.binary(c(1, 0)) + 3 # not binary anymore

try_binary(c(0, 1))
try_binary(c(2, 3))
try_binary(c("a", "b"))

}
