% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\alias{confusion_matrix.default}
\title{Confusion Matrix Metrics}
\usage{
confusion_matrix(data, ...)

\method{confusion_matrix}{default}(data, truth, estimate, na.rm = getOption("na.rm", FALSE), ...)
}
\arguments{
\item{data}{Either a \code{data.frame} containing the columns specified by the
\code{truth} and \code{estimate} arguments, or a \code{table}/\code{matrix} where the true
class results should be in the columns of the table.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{factor} vector.}

\item{estimate}{The column identifier for the predicted class
results (that is also \code{factor}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{factor} vector.}

\item{na.rm}{A \link{logical} to indicate whether missing values should be removed.}
}
\description{
Create a confusion matrix and calculate compatible \code{yardstick} metrics.
}
\details{
This is a function-agnostic \code{yardstick} wrapper: it discovers metric functions exported by \code{yardstick}, filters them by metric type compatible with the provided data (class / class-probability / numeric), and computes what is applicable.
}
\examples{
df <- data.frame(name = c("Yes", "No"),
                 Yes = c(123, 26),
                 No = c(13, 834))
confusion_matrix(df)

iris |>
  ml_decision_trees(Species, quiet = TRUE) |>
  confusion_matrix()
}
