% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_disease_clusters.R
\name{detect_disease_clusters}
\alias{detect_disease_clusters}
\alias{n_clusters}
\alias{has_clusters}
\alias{has_ongoing_cluster}
\alias{has_cluster_before}
\alias{has_cluster_after}
\title{Detect Disease Clusters}
\usage{
detect_disease_clusters(
  df,
  column_date = NULL,
  column_patientid = NULL,
  based_on_historic_maximum = FALSE,
  period_length_months = 12,
  minimum_cases = 5,
  minimum_days = 0,
  minimum_case_days = 2,
  minimum_case_fraction_in_period = 0.02,
  threshold_percentile = 97.5,
  remove_outliers = TRUE,
  remove_outliers_coefficient = 1.5,
  moving_average_days = 7,
  moving_average_side = "left",
  case_free_days = 14,
  ...
)

n_clusters(x)

has_clusters(x, n = 1)

has_ongoing_cluster(x, dates = Sys.Date() - 1)

has_cluster_before(x, date)

has_cluster_after(x, date)
}
\arguments{
\item{df}{Data set: This must consist of \strong{only positive results}. The minimal data set should include a date column and a patient column. Do not summarize on patient IDs; this will be handled automatically.}

\item{column_date}{Name of the column to use for dates. If left blank, the first date column will be used.}

\item{column_patientid}{Name of the column to use for patient IDs. If left blank, the first column resembling \code{"patient|patid"} will be used.}

\item{based_on_historic_maximum}{A \link{logical} to indicate whether the cluster detection should be based on the maximum of previous years. The default is \code{FALSE}, which uses all historic data points.}

\item{period_length_months}{Number of months per period.}

\item{minimum_cases}{Minimum number of \emph{cases} that a cluster requires to be considered a cluster.}

\item{minimum_days}{Minimum number of \emph{days} that a cluster requires to be considered a cluster.}

\item{minimum_case_days}{Minimum number of \emph{days with cases} that a cluster requires to be considered a cluster.}

\item{minimum_case_fraction_in_period}{Minimum fraction of \emph{cluster cases in a period} that a cluster requires to be considered a cluster.}

\item{threshold_percentile}{Threshold to set.}

\item{remove_outliers}{A \link{logical} to indicate whether outliers should be removed before determining the threshold.}

\item{remove_outliers_coefficient}{Coefficient used for outlier determination.}

\item{moving_average_days}{Number of days to set in \code{\link[=moving_average]{moving_average()}}. Defaults to a whole week (\code{7}).}

\item{moving_average_side}{Side of days to set in \code{\link[=moving_average]{moving_average()}}. Defaults to \code{"left"} for retrospective analysis.}

\item{case_free_days}{Number of days to set in \code{\link[AMR:get_episode]{get_episode()}}.}

\item{...}{not used at the moment}

\item{x}{output of \code{\link[=detect_disease_clusters]{detect_disease_clusters()}}}

\item{n}{number of clusters, defaults to 1}

\item{dates}{date(s) to test whether any of the clusters currently has this date in it, defaults to yesterday.}

\item{date}{date to test whether there are any clusters since or until this date.}
}
\description{
Detect disease clusters with \code{\link[=detect_disease_clusters]{detect_disease_clusters()}}. Use \code{\link[=has_clusters]{has_clusters()}} to return \code{TRUE} or \code{FALSE} based on its output, or employ \code{\link[=format]{format()}} to format the result.
}
\details{
A (disease) cluster is defined as an unusually large aggregation of disease events in time or space (\href{https://www.atsdr.cdc.gov/hec/csem/cluster/docs/clusters.pdf}{ATSDR, 2008}). They are common, particularly in large populations. From a statistical standpoint, it is nearly inevitable that some clusters of chronic diseases will emerge within various communities, be it schools, church groups, social circles, or neighborhoods. Initially, these clusters are often perceived as products of specific, predictable processes rather than random occurrences in a particular location, akin to a coin toss.

Whether a (suspected) cluster corresponds to an actual increase of disease in the area, needs to be assessed by an epidemiologist or biostatistician (\href{https://www.atsdr.cdc.gov/hec/csem/cluster/docs/clusters.pdf}{ATSDR, 2008}).

The function \code{\link[=has_ongoing_cluster]{has_ongoing_cluster()}} returns a \link{logical} vector with the same length as \code{dates}, so \code{dates} can have any length.
}
\examples{
cases <- data.frame(date = sample(seq(as.Date("2015-01-01"),
                                      as.Date("2022-12-31"),
                                      "1 day"),
                                  size = 300),
                    patient = sample(LETTERS, size = 300, replace = TRUE))

# -----------------------------------------------------------

check <- detect_disease_clusters(cases, threshold_percentile = 0.99)

has_clusters(check)
check


check2 <- detect_disease_clusters(cases,
                                  minimum_cases = 1,
                                  threshold_percentile = 0.75)

check2
check2 |> format()

check2 |> n_clusters()
check2 |> has_clusters()
check2 |> has_clusters(n = 15)

check2 |> has_ongoing_cluster("2022-06-01")
check2 |> has_ongoing_cluster(c("2022-06-01", "2022-06-20"))
check2 |> has_cluster_before("2022-06-01")
check2 |> has_cluster_after("2022-06-01")

check2 |> unclass()

# plot the results
# check2 |> plot2()
}
\seealso{
\code{\link[=detect_biomarker_changes]{detect_biomarker_changes()}}
}
