% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{different_means}
\alias{different_means}
\alias{mean_harmonic}
\alias{mean_geometric}
\title{Different Means}
\usage{
mean_harmonic(x, ..., na.rm = getOption("na.rm", FALSE))

mean_geometric(x, ..., na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{numeric vector}

\item{...}{arguments passed on to \code{\link[base:mean]{base::mean()}}}

\item{na.rm}{ignore empty values}
}
\description{
Functions to determine harmonic and geometric means.
}
\details{
The harmonic mean can be expressed as the reciprocal of the arithmetic mean of the reciprocals.

The geometric mean is defined as the nth root of the product of n numbers.
}
