% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_warning.R
\name{early_warning_outbreak}
\alias{early_warning_outbreak}
\title{Early Warning for Outbreaks}
\usage{
early_warning_outbreak(
  df,
  ...,
  group_by = NULL,
  column_date = NULL,
  column_patientid = NULL,
  minimum_ongoing_days = 2,
  threshold_percentile = 97.5,
  remove_outliers = TRUE,
  remove_outliers_coefficient = 1.5,
  moving_average_days = 5,
  moving_average_side = "centre",
  case_free_days = 14
)
}
\arguments{
\item{df}{data set, this must be a data set containing \verb{only positive results}. The most bare-bone data set to use must have at least a date column and patient column. Do not summarise on patient IDs, this will happen automatically.}

\item{...}{filter arguments to set apart the data to compare, e.g., \code{jaar == 2023}}

\item{group_by}{column to use in \code{\link[=group_by]{group_by()}}, e.g., \code{testcode} or \code{BepalingCode}}

\item{column_date}{name of the column to use for dates, will take the first date column if left blank}

\item{column_patientid}{name of the column to use for patient IDs, will take the first column names resembling \code{"patient|patid"} if left blank}

\item{threshold_percentile}{threshold to set}

\item{remove_outliers}{\link{logical} to indicate whether outliers must be removed before threshold determination}

\item{remove_outliers_coefficient}{coefficient to use for outlier determination}

\item{moving_average_days}{number of days to set in \code{\link[=moving_average]{moving_average()}}}

\item{moving_average_side}{side of days to set in \code{\link[=moving_average]{moving_average()}}}

\item{case_free_days}{number of days to set in \code{\link[AMR:get_episode]{get_episode()}}}
}
\description{
Feed in positive cases, and this function will determine potential outbreaks. Use \code{\link[=format]{format()}} to format the output.
}
\examples{
cases <- data.frame(date = sample(seq(as.Date("2018-01-01"),
                                      as.Date("2022-12-31"),
                                      "1 day"),
                                  size = 300),
                    patient = sample(LETTERS, size = 300, replace = TRUE))
check <- early_warning_outbreak(cases,
                                date >= "2022-01-01")

format(check)

check2 <- early_warning_outbreak(cases,
                                date >= "2022-01-01",
                                threshold_percentile = 0.75)

print(check2)


unclass(check2)
}
