% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esbl_tests}
\alias{esbl_tests}
\title{Example Data Set with ESBL Test Outcomes}
\format{
A \link{tibble}/\link{data.frame} with 500 observations and 19 variables:
\itemize{
\item \code{esbl}: Outcome of ESBL test (\code{TRUE} (n = 250), \code{FALSE} (n = 250))
\item \code{genus}: Taxonomic genus of the bacteria (\emph{Citrobacter} (n = 42), \emph{Enterobacter} (n = 30), \emph{Escherichia} (n = 301), \emph{Klebsiella} (n = 70), \emph{Morganella} (n = 24), \emph{Proteus} (n = 33))
\item \code{AMC}: MIC values of amoxicillin/clavulanic acid
\item \code{AMP}: MIC values of ampicillin
\item \code{TZP}: MIC values of piperacillin/tazobactam
\item \code{CXM}: MIC values of cefuroxime
\item \code{FOX}: MIC values of cefoxitin
\item \code{CTX}: MIC values of cefotaxime
\item \code{CAZ}: MIC values of ceftazidime
\item \code{GEN}: MIC values of gentamicin
\item \code{TOB}: MIC values of tobramycin
\item \code{TMP}: MIC values of trimethoprim
\item \code{SXT}: MIC values of trimethoprim/sulfamethoxazole
\item \code{NIT}: MIC values of nitrofurantoin
\item \code{FOS}: MIC values of fosfomycin
\item \code{CIP}: MIC values of ciprofloxacin
\item \code{IPM}: MIC values of imipenem
\item \code{MEM}: MIC values of meropenem
\item \code{COL}: MIC values of colistin
}
}
\usage{
esbl_tests
}
\description{
This data set contains phenotypic test outcomes of the presence of ESBL (Extended Spectrum Beta-Lactamase) genes, with the minimum inhibitory concentrations (MIC, in mg/L) for 17 antibiotic drugs.
}
\examples{
head(esbl_tests)
}
\keyword{datasets}
