% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\alias{is_imputed}
\alias{get_mice}
\title{Impute: Filling Missing Values}
\usage{
impute(
  .data,
  vars = everything(),
  algorithm = "mice",
  m = 10,
  method = NULL,
  FUN = mean,
  info = TRUE,
  ...
)

is_imputed(.data)

get_mice(.data)
}
\arguments{
\item{.data}{data set with missing values to impute}

\item{vars}{variables of \code{.data} that must be imputed, defaults to \code{\link[tidyselect:everything]{everything()}} and supports the \code{tidyselect} language.}

\item{algorithm}{algorithm to use for imputation, must be \code{"mice"} or \code{"single-point"}, see \emph{Details}. For the latter, \code{FUN} must be given.}

\item{m}{number of multiple imputations if using MICE, see \code{\link[mice:mice]{mice::mice()}}. The mean of all imputations will be used as result.}

\item{method}{method to use if using MICE, see \code{\link[mice:mice]{mice::mice()}}}

\item{FUN}{function to use for single-point imputation (directly) or for MICE to summarise the results over all \code{m} iterations}

\item{info}{print info about imputation}

\item{...}{arguments to pass on to \code{\link[mice:mice]{mice::mice()}}}
}
\description{
Imputation is the process of replacing missing data with substituted values. This is done because of three main problems that missing data causes: missing data can introduce a substantial amount of bias, make the handling and analysis of the data more arduous, and create reductions in efficiency.
}
\details{
Imputation can be done using single-point, such as the mean or the median, or using \link[mice:mice]{Multivariate Imputations by Chained Equations (MICE)}. Using MICE is a lot more reliable, but also a lot slower, than single-point imputation.

The suggested and default method is MICE. The generated MICE object will be stored as an \link[=attributes]{attribute} with the data, and can be retrieved with \code{\link[=get_mice]{get_mice()}}, containing all specifics about the imputation. MICE is also known as \emph{fully conditional specification} and \emph{sequential regression multiple imputation}. It was designed for data with randomly missing values, though there is simulation evidence to suggest that with a sufficient number of auxiliary variables it can also work on data that are missing not at random.

Use \code{\link[=is_imputed]{is_imputed()}} to get a \link{data.frame} with \code{TRUE}s for all values that were imputed.
}
\examples{
iris2 <- dplyr::as_tibble(iris)
iris2[2, 2] <- NA
iris2[3, 3] <- NA
iris2[4, 5] <- NA
iris
iris2

result <- iris2 |> impute()
result
  
iris2 |> impute(algorithm = "single-point")
iris2 |>
  impute(vars = starts_with("Sepal"),
         algorithm = "single-point")
iris2 |>
  impute(vars = where(is.double),
         algorithm = "single-point",
         FUN = median)
  
result |> is_imputed()
result |> get_mice()
}
