% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{machine_learning}
\alias{machine_learning}
\alias{ml_xg_boost}
\alias{ml_decision_trees}
\alias{ml_random_forest}
\alias{ml_neural_network}
\alias{ml_nearest_neighbour}
\alias{ml_linear_regression}
\alias{ml_logistic_regression}
\alias{confusion_matrix.certestats_ml}
\alias{predict.certestats_ml}
\alias{apply_model_to}
\alias{feature_importances}
\alias{feature_importance_plot}
\alias{roc_plot}
\alias{gain_plot}
\alias{tree_plot}
\alias{correlation_plot}
\alias{get_metrics}
\alias{get_accuracy}
\alias{get_kappa}
\alias{get_recipe}
\alias{get_specification}
\alias{get_rows_testing}
\alias{get_rows_training}
\alias{get_original_data}
\alias{get_roc_data}
\alias{get_coefficients}
\alias{get_model_variables}
\alias{get_variable_weights}
\alias{tune_parameters}
\alias{check_testing_predictions}
\alias{autoplot.certestats_ml}
\alias{autoplot.certestats_feature_importances}
\alias{autoplot.certestats_tuning}
\title{Create a Machine Learning (ML) Model}
\usage{
ml_xg_boost(
  .data,
  outcome,
  predictors = everything(),
  training_fraction = 0.75,
  strata = NULL,
  na_threshold = 0.01,
  correlation_threshold = 0.9,
  centre = FALSE,
  scale = FALSE,
  engine = "xgboost",
  mode = c("classification", "regression", "unknown"),
  trees = 15,
  ...,
  quiet = FALSE
)

ml_decision_trees(
  .data,
  outcome,
  predictors = everything(),
  training_fraction = 0.75,
  strata = NULL,
  na_threshold = 0.01,
  correlation_threshold = 0.9,
  centre = FALSE,
  scale = FALSE,
  engine = "rpart",
  mode = c("classification", "regression", "unknown"),
  tree_depth = 30,
  ...,
  quiet = FALSE
)

ml_random_forest(
  .data,
  outcome,
  predictors = everything(),
  training_fraction = 0.75,
  strata = NULL,
  na_threshold = 0.01,
  correlation_threshold = 0.9,
  centre = FALSE,
  scale = FALSE,
  engine = "ranger",
  mode = c("classification", "regression", "unknown"),
  trees = 500,
  ...,
  quiet = FALSE
)

ml_neural_network(
  .data,
  outcome,
  predictors = everything(),
  training_fraction = 0.75,
  strata = NULL,
  na_threshold = 0.01,
  correlation_threshold = 0.9,
  centre = TRUE,
  scale = TRUE,
  engine = "nnet",
  mode = c("classification", "regression", "unknown"),
  penalty = 0,
  epochs = 100,
  ...,
  quiet = FALSE
)

ml_nearest_neighbour(
  .data,
  outcome,
  predictors = everything(),
  training_fraction = 0.75,
  strata = NULL,
  na_threshold = 0.01,
  correlation_threshold = 0.9,
  centre = TRUE,
  scale = TRUE,
  engine = "kknn",
  mode = c("classification", "regression", "unknown"),
  neighbors = 5,
  weight_func = "triangular",
  ...,
  quiet = FALSE
)

ml_linear_regression(
  .data,
  outcome,
  predictors = everything(),
  training_fraction = 0.75,
  strata = NULL,
  na_threshold = 0.01,
  correlation_threshold = 0.9,
  centre = TRUE,
  scale = TRUE,
  engine = "lm",
  mode = "regression",
  ...,
  quiet = FALSE
)

ml_logistic_regression(
  .data,
  outcome,
  predictors = everything(),
  training_fraction = 0.75,
  strata = NULL,
  na_threshold = 0.01,
  correlation_threshold = 0.9,
  centre = TRUE,
  scale = TRUE,
  engine = "glm",
  mode = "classification",
  penalty = 0.1,
  ...,
  quiet = FALSE
)

\method{confusion_matrix}{certestats_ml}(data, ...)

\method{predict}{certestats_ml}(object, new_data, type = NULL, ...)

apply_model_to(
  object,
  new_data,
  only_prediction = FALSE,
  only_certainty = FALSE,
  correct_mistakes = TRUE,
  impute_algorithm = "mice",
  ...
)

feature_importances(object, ...)

feature_importance_plot(object, ...)

roc_plot(object, ...)

gain_plot(object, ...)

tree_plot(object, ...)

correlation_plot(
  data,
  add_values = TRUE,
  cols = everything(),
  correlation_threshold = 0.9
)

get_metrics(object)

get_accuracy(object)

get_kappa(object)

get_recipe(object)

get_specification(object)

get_rows_testing(object)

get_rows_training(object)

get_original_data(object)

get_roc_data(object)

get_coefficients(object)

get_model_variables(object)

get_variable_weights(object)

tune_parameters(object, ..., only_params_in_model = FALSE, levels = 5, k = 10)

check_testing_predictions(object)

\method{autoplot}{certestats_ml}(object, plot_type = "roc", ...)

\method{autoplot}{certestats_feature_importances}(object, ...)

\method{autoplot}{certestats_tuning}(object, type = c("marginals", "parameters", "performance"), ...)
}
\arguments{
\item{.data}{A data set used for model training and internal testing.
The data will be split into training and testing subsets using
\code{\link[rsample:initial_split]{rsample::initial_split()}}.}

\item{outcome}{Outcome variable, also called the \emph{response} or \emph{dependent}
variable; the variable to be predicted. The value is evaluated using
\code{\link[dplyr:select]{dplyr::select()}} and therefore supports the \code{tidyselect} language.

For classification models, the outcome will be coerced to a \link{factor}
if it is not already a factor or character.}

\item{predictors}{Explanatory variables, also called the \emph{predictors} or
\emph{independent variables}; the variables used to predict \code{outcome}.
This argument supports the \code{tidyselect} language and defaults to
\code{\link[tidyselect:everything]{everything()}}.

Predictors are processed via a \code{\link[recipes:recipe]{recipes::recipe()}}, including dummy
encoding for nominal variables and optional centring, scaling, and
correlation filtering.}

\item{training_fraction}{Fraction of rows to be used for \emph{training},
defaults to \code{0.75}. The remaining rows are used for \emph{testing}.

If a value greater than \code{1} is supplied, it is interpreted as the
absolute number of rows to include in the training set.

When \code{strata} is supplied, the split is stratified accordingly.}

\item{strata}{A variable in \code{data} (single character or name) used to conduct
stratified sampling. When not \code{NULL}, each resample is created within the
stratification variable. Numeric \code{strata} are binned into quartiles.}

\item{na_threshold}{Maximum fraction of \code{NA} values (defaults to \code{0.01}) of the \code{predictors} before they are removed from the model, using \code{\link[recipes:step_rm]{recipes::step_rm()}}}

\item{correlation_threshold}{A value (default 0.9) to indicate the correlation threshold. Predictors with a correlation higher than this value with be removed from the model, using \code{\link[recipes:step_corr]{recipes::step_corr()}}}

\item{centre}{A \link{logical} to indicate whether the \code{predictors} should be transformed so that their mean will be \code{0}, using \code{\link[recipes:step_center]{recipes::step_center()}}. Binary columns will be skipped.}

\item{scale}{A \link{logical} to indicate whether the \code{predictors} should be transformed so that their standard deviation will be \code{1}, using \code{\link[recipes:step_scale]{recipes::step_scale()}}. Binary columns will be skipped.}

\item{engine}{\R package or function name to be used for the model, will be passed on to \code{\link[parsnip:set_engine]{parsnip::set_engine()}}}

\item{mode}{Type of predicted value.
One of \code{"classification"}, \code{"regression"}, or \code{"unknown"}.

If \code{"unknown"}, the mode will be inferred by the underlying \code{parsnip}
model where possible. Explicitly setting \code{mode} is recommended to avoid
ambiguity.}

\item{trees}{An integer for the number of trees contained in
the ensemble.}

\item{...}{Additional arguments passed to the underlying \code{parsnip}
model specification (e.g. \code{trees}, \code{mtry}, \code{penalty}).

For \code{\link[=tune_parameters]{tune_parameters()}}, these must be \code{dials} parameter objects
such as \code{dials::trees()} or \code{dials::mtry()}.

For \code{\link[parsnip:predict.model_fit]{predict()}}, these arguments are
forwarded to \code{\link[parsnip:predict.model_fit]{parsnip::predict.model_fit()}}.
Also see \emph{Model Functions}.}

\item{quiet}{A \link{logical} to silence console output.

For the \code{\link[=tune_parameters]{tune_parameters()}} function, these must be \code{dials} package calls, such as \code{dials::trees()} (see Examples).

For \code{\link[parsnip:predict.model_fit]{predict()}}, these must be arguments passed on to \code{\link[parsnip:predict.model_fit]{parsnip::predict.model_fit()}}}

\item{tree_depth}{An integer for maximum depth of the tree.}

\item{penalty}{A non-negative number representing the total
amount of regularization (specific engines only).}

\item{epochs}{An integer for the number of training iterations.}

\item{neighbors}{A single integer for the number of neighbors
to consider (often called \code{k}). For \pkg{kknn}, a value of 5
is used if \code{neighbors} is not specified.}

\item{weight_func}{A \emph{single} character for the type of kernel function used
to weight distances between samples. Valid choices are: \code{"rectangular"},
\code{"triangular"}, \code{"epanechnikov"}, \code{"biweight"}, \code{"triweight"},
\code{"cos"}, \code{"inv"}, \code{"gaussian"}, \code{"rank"}, or \code{"optimal"}.}

\item{object, data}{outcome of machine learning model}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{type}{A single character value or \code{NULL}. Possible values
are \code{"numeric"}, \code{"class"}, \code{"prob"}, \code{"conf_int"}, \code{"pred_int"},
\code{"quantile"}, \code{"time"}, \code{"hazard"}, \code{"survival"}, or \code{"raw"}. When \code{NULL},
\code{predict()} will choose an appropriate value based on the model's mode.}

\item{only_prediction}{a \link{logical} to indicate whether predictions must be returned as \link{vector}, otherwise returns a \link{data.frame}}

\item{only_certainty}{a \link{logical} to indicate whether certainties must be returned as \link{vector}, otherwise returns a \link{data.frame}}

\item{correct_mistakes}{a \link{logical} to indicate whether missing variables and missing values should be added to \code{new_data}}

\item{impute_algorithm}{the algorithm to use in \code{\link[=impute]{impute()}} if \code{correct_mistakes = TRUE}. Can be \code{"mice"} (default) for the \link[mice:mice]{Multivariate Imputations by Chained Equations (MICE) algorithm}, or \code{"single-point"} for a trained median.}

\item{add_values}{a \link{logical} to indicate whether values must be printed in the tiles}

\item{cols}{columns to use for correlation plot, defaults to \code{\link[dplyr:reexports]{everything()}}}

\item{only_params_in_model}{a \link{logical} to indicate whether only parameters in the model should be tuned}

\item{levels}{An integer for the number of values of each parameter to use
to make the regular grid. \code{levels} can be a single integer or a vector of
integers that is the same length as the number of parameters in \code{...}.
\code{levels} can be a named integer vector, with names that match the id values
of parameters.}

\item{k}{The number of partitions of the data set}

\item{plot_type}{the plot type, can be \code{"roc"} (default), \code{"gain"}, \code{"lift"} or \code{"pr"}. These functions rely on \code{\link[yardstick:roc_curve]{yardstick::roc_curve()}}, \code{\link[yardstick:gain_curve]{yardstick::gain_curve()}}, \code{\link[yardstick:lift_curve]{yardstick::lift_curve()}} and \code{\link[yardstick:pr_curve]{yardstick::pr_curve()}} to construct the curves.}
}
\value{
A trained machine learning model of class \code{certestats_ml}, extending
a \code{\link[parsnip:model_fit]{parsnip::model_fit}} object.

The model object contains additional attributes with training data,
preprocessing steps, predictions, and performance metrics.
}
\description{
These functions create machine learning models using different \code{parsnip}
model specifications and engines, and provide a unified interface for
training, evaluating, tuning, and applying those models.

The \verb{ml_*()} functions are structured wrappers around the \code{tidymodels}
ecosystem, most notably \href{https://parsnip.tidymodels.org}{\code{parsnip}},
\href{https://recipes.tidymodels.org}{\code{recipes}}, \href{https://rsample.tidymodels.org}{\code{rsample}},
\href{https://tune.tidymodels.org}{\code{tune}}, and \href{https://yardstick.tidymodels.org}{\code{yardstick}}.

The goal is to reduce boilerplate while preserving access to the underlying
modeling assumptions, parameters, and diagnostics.
}
\details{
To predict \strong{regression} (numeric values), the function \code{\link[=ml_logistic_regression]{ml_logistic_regression()}} cannot be used.

To predict \strong{classifications} (character values), the function \code{\link[=ml_linear_regression]{ml_linear_regression()}} cannot be used.

The \code{\link[=apply_model_to]{apply_model_to()}} function prioritises successful prediction over strict validation, as \code{correct_mistakes} defaults to \code{TRUE}.

The workflow of the \verb{ml_*()} functions is approximately:

\preformatted{
                       .data
                         |
               rsample::initial_split()
                     /        \
     rsample::training() rsample::testing()
             |                |
     recipes::recipe()        |
             |                |
      recipes::step_corr()    |
             |                |
     recipes::step_center()   |
             |                |
      recipes::step_scale()   |
             |                |
        recipes::prep()       |
         /           \        |
recipes::bake()   recipes::bake()
       |                      |
generics::fit()      yardstick::metrics()
       |                      |
    model object      attributes(model)
}

The \code{\link[parsnip:predict.model_fit]{predict()}} function can be used to fit a model on a new data set. Its wrapper \code{\link[=apply_model_to]{apply_model_to()}} works in the same way, but can also detects and fixes missing variables, missing data points, and data type differences between the trained data and the input data.

Use \code{\link[=feature_importances]{feature_importances()}} to get the importance of all features/variables. Use \code{\link[=autoplot]{autoplot()}} afterwards to plot the results. These two functions are combined in \code{\link[=feature_importance_plot]{feature_importance_plot()}}. Feature importance values are model-specific heuristics and are not comparable across model types.

Use The \code{\link[=tree_plot]{tree_plot()}} to plot the decision tree. For XGBoost models, \code{\link[xgboost:xgb.plot.tree]{xgboost::xgb.plot.tree()}} will be used. For all other tree models, \code{\link[rpart.plot:rpart.plot]{rpart.plot::rpart.plot()}} will be used.

Use \code{\link[=correlation_plot]{correlation_plot()}} to plot the correlation between all variables, even characters. If the input is a \code{certestats} ML model, the training data of the model will be plotted.

Use the \code{\link[=get_model_variables]{get_model_variables()}} function to return a zero-row \link{data.frame} with the variables that were used for training, even before the recipe steps.

Use the \code{\link[=get_variable_weights]{get_variable_weights()}} function to determine the (rough) estimated weights of each variable in the model. This is not as reliable as retrieving coefficients, but it does work for any model. The weights are determined by running the model over all the highest and lowest values of each variable in the trained data. The function returns a data set with 1 row, of which the values sum up to 1.

Use the \code{\link[=tune_parameters]{tune_parameters()}} function to analyse tune parameters of any \verb{ml_*()} function. Without any parameters manually defined, it will try to tune all parameters of the underlying ML model. The tuning will be based on a \link[rsample:vfold_cv]{K-fold cross-validation}, of which the number of partitions can be set with \code{k}. The number of \code{levels} will be used to split the range of the parameters. For example, a range of 1-10 with \code{levels = 2} will lead to \verb{[1, 10]}, while \code{levels = 5} will lead to \verb{[1, 3, 5, 7, 9]}. The resulting \link{data.frame} will be sorted from best to worst. These results can also be plotted using \code{\link[=autoplot]{autoplot()}}.

The \code{\link[=check_testing_predictions]{check_testing_predictions()}} function combines the data used for testing from the original data with its predictions, so the original data can be reviewed per prediction.

Use \code{\link[=autoplot]{autoplot()}} on a model to plot the receiver operating characteristic (ROC) curve, the gain curve, the lift curve, or the precision-recall (PR) curve. For the ROC curve, the (overall) area under the curve (AUC) will be printed as subtitle.
}
\section{Attributes}{

The \verb{ml_*()} functions return a model object with the following
\link[base:attributes]{attributes}:
\itemize{
\item \code{properties}: a \link{list} containing model metadata, including the ML
function used, engine, mode, training/testing sizes, and model-specific
parameters (e.g. \code{trees}, \code{tree_depth})
\item \code{recipe}: a prepped \code{\link[recipes:recipe]{recipes::recipe()}} used for training and testing
\item \code{data_original}: the original input data (after removal of invalid strata)
\item \code{data_structure}: a zero-row \link{data.frame} containing the trained variables
\item \code{data_means}: column means of numeric training variables
\item \code{data_training}: processed training data after recipe baking
\item \code{data_testing}: processed testing data after recipe baking
\item \code{rows_training}: integer indices of training rows in \code{data_original}
\item \code{rows_testing}: integer indices of testing rows in \code{data_original}
\item \code{predictions}: predictions on the testing data
\item \code{metrics}: performance metrics returned by \code{\link[yardstick:metrics]{yardstick::metrics()}}
\item \code{correlation_threshold}: numeric correlation threshold used in preprocessing
\item \code{centre}: logical indicating whether centring was applied
\item \code{scale}: logical indicating whether scaling was applied
}
}

\section{Model Functions}{

These functions wrap \code{parsnip} model specifications. Arguments set in \code{...} will be passed on to these \code{parsnip} functions.
\itemize{
\item \code{ml_decision_trees}: \code{\link[parsnip:decision_tree]{parsnip::decision_tree()}}
\item \code{ml_linear_regression}: \code{\link[parsnip:linear_reg]{parsnip::linear_reg()}}
\item \code{ml_logistic_regression}: \code{\link[parsnip:logistic_reg]{parsnip::logistic_reg()}}
\item \code{ml_neural_network}: \code{\link[parsnip:mlp]{parsnip::mlp()}}
\item \code{ml_nearest_neighbour}: \code{\link[parsnip:nearest_neighbor]{parsnip::nearest_neighbor()}}
\item \code{ml_random_forest}: \code{\link[parsnip:rand_forest]{parsnip::rand_forest()}}
\item \code{ml_xg_boost}: \code{\link[parsnip:boost_tree]{parsnip::boost_tree()}}
}
}

\examples{
# 'esbl_tests' is an included data set, see ?esbl_tests
print(esbl_tests, n = 5)

esbl_tests |> correlation_plot(add_values = FALSE) # red will be removed from model

# predict ESBL test outcome based on MICs using 2 different models
model1 <- esbl_tests |> ml_xg_boost(esbl, where(is.double))
model2 <- esbl_tests |> ml_decision_trees(esbl, where(is.double))


# Assessing A Model ----------------------------------------------------

model1 |> get_metrics()
model2 |> get_metrics()

model1 |> confusion_matrix()

# a correlation plot of a model shows the training data
model1 |> correlation_plot(add_values = FALSE)

model1 |> feature_importances()
model1 |> feature_importances() |> autoplot()
model2 |> feature_importance_plot()

# decision trees can also have a tree plot
model2 |> tree_plot()


# Applying A Model -----------------------------------------------------
 
# simply use base R `predict()` to apply a model:
model1 |> predict(esbl_tests)

# but apply_model_to() contains more info and can apply corrections:
model1 |> apply_model_to(esbl_tests)
# and to format the result, e.g. for an API:
model1 |> apply_model_to(esbl_tests) |> dplyr::slice(1:10) |> format()

# put in only parts of new data:
model1 |> apply_model_to(esbl_tests[, 1:15])
esbl_tests2 <- esbl_tests
esbl_tests2[2, "CIP"] <- NA
esbl_tests2[5, "AMC"] <- NA
# with XGBoost, nothing will be changed (it can correct for missings):
model1 |> apply_model_to(esbl_tests2)
# with random forest (or others), missings will be imputed:
model2 |> apply_model_to(esbl_tests2)


# Tuning A Model -------------------------------------------------------
 
# tune the parameters of a model (will take some time)
tuning <- model2 |> 
  tune_parameters(k = 5, levels = 3)
autoplot(tuning)

# tuning analysis by specifying (some) parameters
iris |> 
  ml_xg_boost(Species) |> 
  tune_parameters(mtry = dials::mtry(range = c(1, 3)),
                  trees = dials::trees())


# Practical Example #1 --------------------------------------------------

# this is what iris data set looks like:
head(iris)
# create a model to predict the species:
iris_model <- iris |> ml_xg_boost(Species)
iris_model_rf <- iris |> ml_random_forest(Species)
# is it a bit reliable?
get_metrics(iris_model)

# now try to predict species from an arbitrary data set:
to_predict <- data.frame(Sepal.Length = 5,
                         Sepal.Width = 3,
                         Petal.Length = 1.5,
                         Petal.Width = 0.5)
to_predict

# should be 'setosa' in the 'predicted' column with high certainty:
iris_model |> apply_model_to(to_predict)

# API formatting:
iris_model |> apply_model_to(to_predict) |> format()
iris_model |> apply_model_to(to_predict, only_prediction = TRUE)
iris_model |> apply_model_to(to_predict, only_certainty = TRUE)

# which variables are generally important (only trained variables)?
iris_model |> feature_importances()

# how would the model do without the 'Sepal.Length' column?
to_predict <- to_predict[, c("Sepal.Width", "Petal.Width", "Petal.Length")]
to_predict
iris_model |> apply_model_to(to_predict)

# now compare that with a random forest model that requires imputation:
iris_model_rf |> apply_model_to(to_predict)

# the certainly is very different.


# Practical Example #2 -------------------------------------------------

# this example shows plotting methods for a model

# train model to predict genus based on MICs:
genus <- esbl_tests |> ml_xg_boost(genus, everything())
genus |> get_metrics()
genus |> confusion_matrix()
genus |> feature_importance_plot()
genus |> autoplot()
genus |> autoplot(plot_type = "gain")
genus |> autoplot(plot_type = "pr")
}
