% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_from_R_with_global_na.rm.R
\name{math_functions}
\alias{math_functions}
\alias{any}
\alias{all}
\alias{mean}
\alias{sum}
\alias{prod}
\alias{min}
\alias{max}
\alias{pmin}
\alias{pmax}
\alias{range}
\alias{sd}
\alias{var}
\alias{median}
\alias{fivenum}
\alias{quantile}
\alias{IQR}
\title{Mathematical Functions With Global \code{na.rm}}
\source{
\code{\link[base:any]{base::any()}}

\code{\link[base:all]{base::all()}}

\code{\link[base:mean]{base::mean()}}

\code{\link[base:sum]{base::sum()}}

\code{\link[base:prod]{base::prod()}}

\code{\link[base:Extremes]{base::min()}}

\code{\link[base:Extremes]{base::max()}}

\code{\link[base:Extremes]{base::pmin()}}

\code{\link[base:Extremes]{base::pmax()}}

\code{\link[base:range]{base::range()}}

\code{\link[stats:sd]{stats::sd()}}

\code{\link[stats:cor]{stats::var()}}

\code{\link[stats:median]{stats::median()}}

\code{\link[stats:fivenum]{stats::fivenum()}}

\code{\link[stats:quantile]{stats::quantile()}}

\code{\link[stats:IQR]{stats::IQR()}}
}
\usage{
any(..., na.rm = getOption("na.rm", FALSE))

all(..., na.rm = getOption("na.rm", FALSE))

mean(x, ..., na.rm = getOption("na.rm", FALSE))

sum(..., na.rm = getOption("na.rm", FALSE))

prod(..., na.rm = getOption("na.rm", FALSE))

min(..., na.rm = getOption("na.rm", FALSE))

max(..., na.rm = getOption("na.rm", FALSE))

pmin(..., na.rm = getOption("na.rm", FALSE))

pmax(..., na.rm = getOption("na.rm", FALSE))

range(..., na.rm = getOption("na.rm", FALSE))

sd(x, na.rm = getOption("na.rm", FALSE))

var(x, ..., na.rm = getOption("na.rm", FALSE))

median(x, na.rm = getOption("na.rm", FALSE), ...)

fivenum(x, na.rm = getOption("na.rm", FALSE))

quantile(
  x,
  ...,
  na.rm = getOption("na.rm", FALSE),
  type = getOption("quantile.type", 7)
)

IQR(
  x,
  ...,
  na.rm = getOption("na.rm", FALSE),
  type = getOption("quantile.type", 7)
)
}
\arguments{
\item{...}{zero or more logical vectors.  Other objects of zero
    length are ignored, and the rest are coerced to logical ignoring
    any class.}

\item{na.rm}{logical.  If true \code{NA} values are removed before
    the result is computed.}

\item{x}{An \R object.  Currently there are methods for
    numeric/logical vectors and \link[base:Dates]{date},
    \link[base]{date-time} and \link[base]{time interval} objects.  Complex vectors
    are allowed for \code{trim = 0}, only.}

\item{type}{an integer between 1 and 9 selecting one of the
    nine quantile algorithms detailed below to be used.}
}
\description{
These functions call their original base \R namesake, but with a global settable \code{na.rm} argument.
}
\section{Default values of \code{na.rm}}{

This 'certestats' package supports a global default setting for \code{na.rm} in many mathematical functions. This can be set with \code{options(na.rm = TRUE)} or \code{options(na.rm = FALSE)}.

For \code{\link[=normality]{normality()}}, \code{\link[=quantile]{quantile()}} and \code{\link[=IQR]{IQR()}}, this also applies to the \code{type} argument. The default, \code{type = 7}, is the default of base \R. Use \code{type = 6} to comply with SPSS.
}

