% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\alias{moving_sum}
\alias{moving_Q1}
\alias{moving_Q3}
\alias{moving_fn}
\title{Moving Average}
\usage{
moving_average(x, w, side = "centre", na.rm = getOption("na.rm", FALSE))

moving_sum(x, w, side = "centre", na.rm = getOption("na.rm", FALSE))

moving_Q1(x, w, side = "centre", na.rm = getOption("na.rm", FALSE))

moving_Q3(x, w, side = "centre", na.rm = getOption("na.rm", FALSE))

moving_fn(x, w, fun, side = "centre", ...)
}
\arguments{
\item{x}{vector of values}

\item{w}{window length; total number of observations to include. This should preferably be an odd number, so that the same number of values to the left and right of \code{x} are included.}

\item{side}{default is \code{"centre"}, can also be \code{"left"} or \code{"right"}. This can be used to take a moving average (or sum, or ...) of e.g. the last 7 days.}

\item{na.rm}{a \link{logical} to indicate whether empty must be removed from \code{x}}

\item{fun}{function to apply}

\item{...}{arguments passed on to \code{fun}}
}
\description{
Functions to calculate a moving average. These are useful to get rid of (strong) peakiness.
}
\details{
Each function can be used over a moving period with \code{\link[=moving_fn]{moving_fn()}}. For example, for a moving median: \code{moving_fn(x, 7, fun = median)}. Or a moving maximum: \code{moving_fn(x, 5, fun = max)}.

The moving average is determined by averaging \code{floor(w / 2)} values before and after each element of \code{x} and all elements in between.
}
\section{Default values of \code{na.rm}}{

This 'certestats' package supports a global default setting for \code{na.rm} in many mathematical functions. This can be set with \code{options(na.rm = TRUE)} or \code{options(na.rm = FALSE)}.

For \code{\link[=normality]{normality()}}, \code{\link[=quantile]{quantile()}} and \code{\link[=IQR]{IQR()}}, this also applies to the \code{type} argument. The default, \code{type = 7}, is the default of base \R. Use \code{type = 6} to comply with SPSS.
}

