% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson.R
\name{nelson_rules}
\alias{nelson_rules}
\alias{nelson_rule1}
\alias{nelson_rule2}
\alias{nelson_rule3}
\alias{nelson_rule4}
\alias{nelson_rule5}
\alias{nelson_rule6}
\alias{nelson_rule7}
\alias{nelson_rule8}
\alias{nelson_text}
\alias{qc_test}
\title{Nelson's Quality Control Rules}
\source{
Nelson LS. \strong{The Shewhart Control Chart—Tests for Special Causes}. Journal of Quality Technology \link{Internet}. Informa UK Limited; 1984 Oct;16(4):237–9. Available from: \url{http://dx.doi.org/10.1080/00224065.1984.11978921}
}
\usage{
nelson_rule1(x, m = mean(x), s = sd(x))

nelson_rule2(x, m = mean(x), threshold = 9)

nelson_rule3(x, threshold = 6)

nelson_rule4(x, m = mean(x), threshold = 14, direction_mean = FALSE)

nelson_rule5(x, m = mean(x), s = sd(x), threshold = 3)

nelson_rule6(x, m = mean(x), s = sd(x), threshold = 5)

nelson_rule7(x, m = mean(x), s = sd(x), threshold = 15)

nelson_rule8(x, m = mean(x), s = sd(x), threshold = 8)

nelson_text(rule, threshold)

qc_test(x, m = mean(x), s = sd(x), type = "Nelson")
}
\arguments{
\item{x}{vector with values}

\item{m}{mean}

\item{s}{standard deviation}

\item{threshold}{minimal number of sequential values before rule is triggered (defaults to Nelson's)}

\item{direction_mean}{a logical to indicate whether \emph{n} observations in a row must be tested for alternating in direction of the mean}

\item{rule}{number of the rule}

\item{type}{type of QC rules set, must be \code{"Nelson"}, \code{"Westgard"}, \code{"AIAG"}, \code{"Montgomery"}, or \code{"Healthcare"}}
}
\description{
These rules are used for quality control (QC). Default values are set for Nelson's criteria, but they also support Westgard, AIAG, Montgomery and Healthcare QC rules.
}
\section{Rules list}{
\tabular{clccccc}{
   Rule \tab Rule explanation: \tab Nelson \tab Westgard \tab AIAG \tab Montg. \tab HC \cr
   #1 \tab One point is more than 3 standard deviations from the mean \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \cr
   #2 \tab \emph{n} (or more) points in a row are on the same side of the mean \tab 9 \tab 9 \tab 7 \tab 8 \tab 8 \cr
   #3 \tab \emph{n} (or more) points in a row are continually incr. or decr. \tab 6 \tab - \tab 6 \tab 6 \tab 6 \cr
   #4 \tab \emph{n} (or more) points in a row alternate in direction, incr. then decr. \tab 14 \tab - \tab 14 \tab 14 \tab - \cr
   #5 \tab \emph{n}\code{ - 1} out of \emph{n} points in a row are >2 sd from the mean \tab 3 \tab 3 \tab 3 \tab 3 \tab 3 \cr
   #6 \tab \emph{n}\code{ - 1} out of \emph{n} points in a row are >1 sd from the mean \tab 5 \tab 5 \tab 5 \tab 5 \tab - \cr
   #7 \tab >=\emph{n} points in a row are within 1 sd of the mean \tab 15 \tab - \tab 15 \tab 15 \tab 15 \cr
   #8 \tab >=\emph{n} points in a row outside 1 sd of the mean, in both directions \tab 8 \tab - \tab 8 \tab 8 \tab - \cr
}


\emph{Montg.}: Montgomery, \emph{HC}: Healthcare
}

\examples{
x <- qc_test(rnorm(250))
x

if (require("certeplot2")) {
  plot2(x)
}
}
