% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentiles.R
\name{percentiles}
\alias{percentiles}
\alias{deciles}
\title{Transform Vector to Percentiles}
\usage{
percentiles(x)

deciles(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
These functions take a numeric vector as input, and return the lowest \href{https://en.wikipedia.org/wiki/Percentile}{percentiles} or \href{https://en.wikipedia.org/wiki/Decile}{deciles} for each value.
}
\examples{
x <- c(0, 1, 2, 3, 4, 5, 5, 5, 5, 5, 5, 5, 6)
percentiles(x)
deciles(x)

percentiles(rnorm(10))

library(dplyr, warn.conflicts = FALSE)
tib <- as_tibble(matrix(as.integer(runif(40, min = 1, max = 7)), ncol = 4),
                 .name_repair = function(...) LETTERS[1:4])
tib

# percentiles per column
tib |> mutate_all(percentiles)
}
