% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rules.R
\name{qc_rules}
\alias{qc_rules}
\alias{qc_rule1}
\alias{qc_rule2}
\alias{qc_rule3}
\alias{qc_rule4}
\alias{qc_rule5}
\alias{qc_rule6}
\alias{qc_rule7}
\alias{qc_rule8}
\alias{qc_rule_text}
\alias{qc_test}
\title{Quality Control (QC) Rules}
\source{
Nelson LS. \strong{The Shewhart Control Chart—Tests for Special Causes}. Journal of Quality Technology. Informa UK Limited; 1984 Oct;16(4):237–9. \doi{10.1080/00224065.1984.11978921}.
}
\usage{
qc_rule1(x, m = mean(x), s = sd(x))

qc_rule2(x, m = mean(x), threshold = 9)

qc_rule3(x, threshold = 6)

qc_rule4(x, m = mean(x), threshold = 14, direction_mean = FALSE)

qc_rule5(x, m = mean(x), s = sd(x), threshold = 3)

qc_rule6(x, m = mean(x), s = sd(x), threshold = 5)

qc_rule7(x, m = mean(x), s = sd(x), threshold = 15)

qc_rule8(x, m = mean(x), s = sd(x), threshold = 8)

qc_rule_text(rule, threshold)

qc_test(x, m = mean(x), s = sd(x), guideline = "Nelson")
}
\arguments{
\item{x}{vector with values}

\item{m}{mean}

\item{s}{standard deviation}

\item{threshold}{minimal number of sequential values before rule is triggered (defaults to Nelson's)}

\item{direction_mean}{a logical to indicate whether \emph{n} observations in a row must be tested for alternating in direction of the mean}

\item{rule}{number of the rule}

\item{guideline}{guideline of QC rules set, must be \code{"Nelson"}, \code{"Westgard"}, \code{"AIAG"}, \code{"Montgomery"}, or \code{"Healthcare"}}
}
\description{
These rules are used for quality control (QC). Default values are set for \strong{Nelson's QC rules}, but they also support Westgard, AIAG, Montgomery and Healthcare QC rules.
}
\section{Rules list}{
\tabular{clccccc}{
   Rule \tab Rule explanation: \tab Nelson \tab Westgard \tab AIAG \tab Montg. \tab HC \cr
   #1 \tab One point is more than 3 standard deviations from the mean \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \cr
   #2 \tab \code{n} (or more) points in a row are on the same side of the mean \tab 9 \tab 9 \tab 7 \tab 8 \tab 8 \cr
   #3 \tab \code{n} (or more) points in a row are continually incr. or decr. \tab 6 \tab - \tab 6 \tab 6 \tab 6 \cr
   #4 \tab \code{n} (or more) points in a row alternate in direction, incr. then decr. \tab 14 \tab - \tab 14 \tab 14 \tab - \cr
   #5 \tab \code{n - 1} out of \code{n} points in a row are >2 sd from the mean \tab 3 \tab 3 \tab 3 \tab 3 \tab 3 \cr
   #6 \tab \code{n - 1} out of \code{n} points in a row are >1 sd from the mean \tab 5 \tab 5 \tab 5 \tab 5 \tab - \cr
   #7 \tab >=\code{n} points in a row are within 1 sd of the mean \tab 15 \tab - \tab 15 \tab 15 \tab 15 \cr
   #8 \tab >=\code{n} points in a row outside 1 sd of the mean, in both directions \tab 8 \tab - \tab 8 \tab 8 \tab - \cr
}


\emph{Montg.}: Montgomery, \emph{HC}: Healthcare
}

\examples{
x <- qc_test(rnorm(250))
x

# turn into data.frame, e.g. for export
head(as.data.frame(x))

if (require("certeplot2")) {
  plot2(x,
        subtitle = "Workflow 'example123'")
}
}
