% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{everything}
\alias{confusionMatrix}
\alias{metrics}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{caret}{\code{\link[caret]{confusionMatrix}}}

  \item{dplyr}{\code{\link[dplyr:reexports]{everything}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{yardstick}{\code{\link[yardstick]{metrics}}}
}}

