% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove Outliers}
\usage{
remove_outliers(x, coef = 1.5)
}
\arguments{
\item{x}{a vector of values}

\item{coef}{a multiple of the IQR that is allowed at maximum to keep values within the accepted range}
}
\description{
This simple function uses \code{\link[=boxplot.stats]{boxplot.stats()}} to determine outliers and removes them from the vector.
}
\examples{
remove_outliers(c(1,2,1,2,1,2,8))

remove_outliers(c(1,2,1,2,1,2))
}
