% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{row_function}
\alias{row_function}
\title{Apply Function per Row}
\usage{
row_function(fn, ..., data = NULL)
}
\arguments{
\item{fn}{function to apply, such as \code{\link[=max]{max()}}}

\item{...}{tidyverse selector helpers, passed on to \code{\link[dplyr:select]{select()}}}

\item{data}{data set, will be determined with \code{\link[dplyr:deprec-context]{cur_data_all()}} if left blank}
}
\description{
This can be used to e.g. add a maximum of certain rows.
}
\examples{
if (require("dplyr")) {
  iris |> 
    mutate(max = row_function(max, where(is.numeric)),
           sepal_mean = row_function(mean, starts_with("Sepal"))) |> 
    head()
}
}
