% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certepackage.R
\docType{package}
\name{certestats-package}
\alias{certestats}
\alias{certestats-package}
\title{certestats: A Certe R Package for Statistical Modelling}
\description{
A Certe R Package for applying statistical modelling (such as creating machine learning models), QC rules and distribution analysis. This package is part of the 'certedata' universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://certe-medical-epidemiology.github.io/certestats}
  \item \url{https://github.com/certe-medical-epidemiology/certestats}
}

}
\author{
\strong{Maintainer}: Erwin Dijkstra \email{e.dijkstra@certe.nl}

Authors:
\itemize{
  \item Matthijs S. Berends \email{m.berends@certe.nl}
}

Other contributors:
\itemize{
  \item Certe Medical Diagnostics & Advice Foundation [copyright holder, funder]
}

}
\keyword{internal}
