% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_metrics.R
\name{distribution_metrics}
\alias{distribution_metrics}
\alias{se}
\alias{cv}
\alias{cqv}
\alias{z_score}
\alias{midhinge}
\title{Distribution Metrics}
\usage{
se(x, na.rm = getOption("na.rm", FALSE))

cv(x, na.rm = getOption("na.rm", FALSE))

cqv(x, na.rm = getOption("na.rm", FALSE))

z_score(x, na.rm = getOption("na.rm", FALSE))

midhinge(x, na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{value}

\item{na.rm}{a \link{logical} to indicate whether empty must be removed from \code{x}}
}
\description{
These are simple distribution metric functions.
}
\details{
These are the functions:
\itemize{
\item \code{\link[=se]{se()}} (standard error): sd / square root of length
}

\itemize{
\item \code{\link[=cv]{cv()}} (coefficient of variation): standard deviation / mean
}

\itemize{
\item \code{\link[=cqv]{cqv()}} (coefficient of quartile variation): (Q3 - Q1) / (Q3 + Q1)
}

\itemize{
\item \code{\link[=z_score]{z_score()}} (number of standard deviations from the mean): (x - mean) / sd
}

\itemize{
\item \code{\link[=midhinge]{midhinge()}} (mean of interquartile range): (Q1 + Q3) / 2
}
}
\section{Default values of \code{na.rm}}{

This package supports a global default setting for \code{na.rm} in many mathematical functions. This can be set with \code{options(na.rm = TRUE)} or \code{options(na.rm = FALSE)}.
}

