% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_metrics.R
\name{distribution_metrics}
\alias{distribution_metrics}
\alias{se}
\alias{cv}
\alias{cqv}
\alias{z_score}
\alias{midhinge}
\title{Distribution Metrics}
\usage{
se(x, na.rm = TRUE)

cv(x, na.rm = TRUE)

cqv(x, na.rm = TRUE)

z_score(x, na.rm = TRUE)

midhinge(x, na.rm = TRUE)
}
\arguments{
\item{x}{value}

\item{na.rm}{a \link{logical} to indicate whether empty must be removed from \code{x}}
}
\description{
These are simple distribution metric functions.
}
\details{
These are the functions:

\itemize{
\item \code{\link[=cv]{cv()}} (coefficient of variation): standard deviation / mean
}

\itemize{
\item \code{\link[=cqv]{cqv()}} (coefficient of quartile variation): (Q3 - Q1) / (Q3 + Q1)
}

\itemize{
\item \code{\link[=z_score]{z_score()}}: (x - mean) / sd
}

\itemize{
\item \code{\link[=midhinge]{midhinge()}}: (Q1 + Q3) / 2
}
}
