% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_metrics.R
\name{distribution_metrics}
\alias{distribution_metrics}
\alias{se}
\alias{ci}
\alias{sum_of_squares}
\alias{cv}
\alias{cqv}
\alias{mse}
\alias{mape}
\alias{rmse}
\alias{mae}
\alias{z_score}
\alias{midhinge}
\alias{ewma}
\alias{rr_ewma}
\alias{normalise}
\alias{normalize}
\alias{scale_sd}
\alias{centre_mean}
\alias{percentiles}
\alias{deciles}
\title{Distribution Metrics}
\usage{
se(x, na.rm = getOption("na.rm", FALSE))

ci(x, level = 0.95, na.rm = getOption("na.rm", FALSE))

sum_of_squares(x, correct_mean = TRUE, na.rm = getOption("na.rm", FALSE))

cv(x, na.rm = getOption("na.rm", FALSE))

cqv(x, na.rm = getOption("na.rm", FALSE))

mse(actual, predicted, na.rm = getOption("na.rm", FALSE))

mape(actual, predicted, na.rm = getOption("na.rm", FALSE))

rmse(actual, predicted, na.rm = getOption("na.rm", FALSE))

mae(actual, predicted, na.rm = getOption("na.rm", FALSE))

z_score(x, na.rm = getOption("na.rm", FALSE))

midhinge(x, na.rm = getOption("na.rm", FALSE))

ewma(x, lambda, na.rm = getOption("na.rm", FALSE))

rr_ewma(x, lambda, na.rm = getOption("na.rm", FALSE))

normalise(n, n_ref, per = 1000)

normalize(n, n_ref, per = 1000)

scale_sd(x)

centre_mean(x)

percentiles(x, na.rm = getOption("na.rm", FALSE))

deciles(x, na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{values}

\item{na.rm}{a \link{logical} to indicate whether empty must be removed from \code{x}}

\item{level}{alpha level, defaults to 95\%}

\item{correct_mean}{with \code{TRUE} (the default) correct for the mean will be applied, by summing each square of \code{x} after the mean of \code{x} has been subtracted, so that this says something about \code{x}. With \code{FALSE}, all \code{x^2} are simply added together, so this says something about \code{x}'s location in the data.}

\item{actual}{Vector of actual values}

\item{predicted}{Vector of predicted values}

\item{lambda}{smoothing parameter, a value between 0 and 1. A value of 0 is equal to \code{x}, a value of 1 equal to the \emph{mean} of \code{x}. The EWMA looks back and has a delay - the rrEWMA takes the mean of a 'forward' and 'backward' EWMA.}

\item{n}{number to be normalised}

\item{n_ref}{reference to use for normalisation}

\item{per}{normalisation factor}
}
\description{
These are simple distribution metric functions.
}
\details{
These are the explanations of the functions:
\itemize{
\item \code{\link[=se]{se()}} calculates the standard error: sd / square root of length
}

\itemize{
\item \code{\link[=ci]{ci()}} calculates the confidence intervals for a mean (defaults at 95\%), which returns length 2
}

\itemize{
\item \code{\link[=sum_of_squares]{sum_of_squares()}} calculates the sum of (x - mean(x)) ^ 2
}

\itemize{
\item \code{\link[=cv]{cv()}} calculates the coefficient of variation: standard deviation / mean
}

\itemize{
\item \code{\link[=cqv]{cqv()}} calculates the coefficient of quartile variation: (Q3 - Q1) / (Q3 + Q1)
}

\itemize{
\item \code{\link[=mse]{mse()}} calculates the mean squared error
}

\itemize{
\item \code{\link[=mape]{mape()}} calculates the mean absolute percentage error
}

\itemize{
\item \code{\link[=rmse]{rmse()}} calculates the root mean squared error
}

\itemize{
\item \code{\link[=mae]{mae()}} calculates the mean absolute error
}

\itemize{
\item \code{\link[=z_score]{z_score()}} calculates the number of standard deviations from the mean: (x - mean) / sd
}

\itemize{
\item \code{\link[=midhinge]{midhinge()}} calculates the mean of interquartile range: (Q1 + Q3) / 2
}

\itemize{
\item \code{\link[=ewma]{ewma()}} calculates the EWMA (exponentially weighted moving average)
}

\itemize{
\item \code{\link[=rr_ewma]{rr_ewma()}} calculates the rrEWMA (reversed-recombined exponentially weighted moving average)
}

\itemize{
\item \code{\link[=normalise]{normalise()}} normalises the data based on a reference: (n / reference) * unit
}

\itemize{
\item \code{\link[=scale_sd]{scale_sd()}} normalises the data to have a standard deviation of 1, while retaining the mean
}

\itemize{
\item \code{\link[=centre_mean]{centre_mean()}} normalises the data to have a mean of 0, while retaining the standard deviation
}

\itemize{
\item \code{\link[=percentiles]{percentiles()}} and \code{\link[=deciles]{deciles()}} take a numeric vector as input, and return the lowest \href{https://en.wikipedia.org/wiki/Percentile}{percentiles} or \href{https://en.wikipedia.org/wiki/Decile}{deciles} for each value
}
}
\section{Default values of \code{na.rm}}{

This 'certestats' package supports a global default setting for \code{na.rm} in many mathematical functions. This can be set with \code{options(na.rm = TRUE)} or \code{options(na.rm = FALSE)}.

For \code{\link[=normality]{normality()}}, \code{\link[=quantile]{quantile()}} and \code{\link[=IQR]{IQR()}}, this also applies to the \code{type} argument. The default, \code{type = 7}, is the default of base \R. Use \code{type = 6} to comply with SPSS.
}

\examples{
x <- c(0, 1, 2, 3, 4, 5, 5, 5, 5, 5, 5, 5, 6)
percentiles(x)
deciles(x)

percentiles(rnorm(10))

library(dplyr, warn.conflicts = FALSE)
tib <- as_tibble(matrix(as.integer(runif(40, min = 1, max = 7)), ncol = 4),
                 .name_repair = function(...) LETTERS[1:4])
tib

# percentiles per column
tib |> mutate_all(percentiles)
}
\seealso{
For the sum of squares: \url{https://www.thoughtco.com/sum-of-squares-formula-shortcut-3126266}
}
