% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\alias{regression.default}
\alias{regression.data.frame}
\alias{plot.certestats_reg}
\alias{autoplot.certestats_reg}
\title{Fast Regression Models}
\usage{
regression(x, ...)

\method{regression}{default}(x, y = NULL, type = "lm", family = stats::gaussian, ...)

\method{regression}{data.frame}(x, var1, var2 = NULL, type = "lm", family = stats::gaussian, ...)

\method{plot}{certestats_reg}(x, ...)

\method{autoplot}{certestats_reg}(object, ...)
}
\arguments{
\item{x}{vector of values, or a data.frame}

\item{...}{arguments for \code{\link[=lm]{lm()}} or \code{\link[=glm]{glm()}}}

\item{y}{vector of values, optional}

\item{type}{type of function to use, can be "lm" or "glm"}

\item{family}{only used for \code{\link[=glm]{glm()}}}

\item{var1, var2}{column to use of \code{x}, the \code{var2} argument is optional}

\item{object}{data to plot}
}
\description{
Functions to do fast regression modelling. The functions return a \link{tibble} with statistics. Use \code{\link[=plot]{plot()}} for an extensive model visualisation.
}
\examples{
runif(10) |> regression()

data.frame(x = 1:50, y = runif(50)) |>
  regression(x, y)

mrsa_from_blood_years <- c(0, 1, 0, 0, 2, 0, 1, 3, 1, 2, 3, 1, 2)
mrsa_from_blood_years |> plot()

mrsa_from_blood_years |> regression()

mrsa_from_blood_years |> regression() |> plot()
}
