% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean.R
\name{weighted_mean}
\alias{weighted_mean}
\alias{weighted_median}
\alias{weighted_Q1}
\alias{weighted_Q3}
\alias{weighted_fn}
\title{Weighted Mean}
\usage{
weighted_mean(x, w, na.rm = getOption("na.rm", FALSE))

weighted_median(x, w, na.rm = getOption("na.rm", FALSE))

weighted_Q1(x, w, na.rm = getOption("na.rm", FALSE))

weighted_Q3(x, w, na.rm = getOption("na.rm", FALSE))

weighted_fn(x, w, fun, ...)
}
\arguments{
\item{x}{vector of values}

\item{w}{weights}

\item{na.rm}{a \link{logical} to indicate whether empty must be removed from \code{x}}

\item{fun}{function to apply}

\item{...}{arguments passed on to \code{fun}}
}
\description{
Functions to calculate a weighted mean.
}
\section{Default values of \code{na.rm}}{

This 'certestats' package supports a global default setting for \code{na.rm} in many mathematical functions. This can be set with \code{options(na.rm = TRUE)} or \code{options(na.rm = FALSE)}.

For \code{\link[=quantile]{quantile()}} and \code{\link[=IQR]{IQR()}}, this also applies to the \code{type} argument. The default, \code{type = 7}, is the default of base \R. Use \code{type = 6} to comply with SPSS.
}

\examples{
x <- c(1:10)
y <- c(1:10)

mean(x)
weighted_mean(x, y)

# essentially equal to:
mean(rep(x, y))
}
